Update tex_projecting_overpaint___script.py
Add functionality to set found node as active before projection

This commit modifies the OverpaintCameraProjection operator to ensure that the node found is set as the active node before executing the projection. If the node is located within a node group, the script also ensures that the found node is set as the active node within its group. Additionally, it cancels any existing selections or activations of nodes before setting the target node as active.